# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vigyanics API — public auth + admin panel
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface CreateAdminInput {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface UserProfile {
  id: number;
  email: string;
  role: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface DashboardStats {
  totalProducts: number;
  activeListings: number;
  totalOrders: number;
  totalRevenue: number;
  totalCustomers: number;
  lowStockItems: number;
  pendingOrders: number;
}

export interface ProductImage {
  id: number;
  productId: number;
  url: string;
  /** @nullable */
  alt?: string | null;
  sortOrder: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  longDescription?: string | null;
  price: string;
  /** @nullable */
  salePrice?: string | null;
  /** @nullable */
  sku?: string | null;
  quantity: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  tags?: string[] | null;
  /** @nullable */
  thumbnail?: string | null;
  status: string;
  stockStatus: string;
  isFeatured: boolean;
  isTrending: boolean;
  isBestSeller: boolean;
  isNewArrival: boolean;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  features?: string[] | null;
  images?: ProductImage[];
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  longDescription?: string | null;
  price: string;
  /** @nullable */
  salePrice?: string | null;
  /** @nullable */
  sku?: string | null;
  quantity?: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  tags?: string[] | null;
  /** @nullable */
  thumbnail?: string | null;
  status?: string;
  stockStatus?: string;
  isFeatured?: boolean;
  isTrending?: boolean;
  isBestSeller?: boolean;
  isNewArrival?: boolean;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  features?: string[] | null;
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  longDescription?: string | null;
  price?: string;
  /** @nullable */
  salePrice?: string | null;
  /** @nullable */
  sku?: string | null;
  quantity?: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  tags?: string[] | null;
  /** @nullable */
  thumbnail?: string | null;
  status?: string;
  stockStatus?: string;
  isFeatured?: boolean;
  isTrending?: boolean;
  isBestSeller?: boolean;
  isNewArrival?: boolean;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  features?: string[] | null;
}

export interface ProductListResponse {
  data: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface BulkIdsInput {
  ids: number[];
}

export interface BulkResult {
  deleted: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
  productCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
}

export interface Order {
  id: number;
  orderNumber: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  status: string;
  total: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  /** @nullable */
  productId?: number | null;
  productName: string;
  quantity: number;
  price: string;
  total: string;
}

/**
 * @nullable
 */
export type OrderDetailShippingAddress = { [key: string]: unknown } | null;

export interface OrderDetail {
  id: number;
  orderNumber: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  status: string;
  total: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  shippingAddress?: OrderDetailShippingAddress;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface OrderListResponse {
  data: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface CustomerListResponse {
  data: UserProfile[];
  total: number;
  page: number;
  limit: number;
}

export interface CustomerStatusUpdate {
  isActive: boolean;
}

export type ListAdminProductsParams = {
search?: string;
status?: string;
categoryId?: number;
page?: number;
limit?: number;
};

export type ListAdminOrdersParams = {
status?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListAdminCustomersParams = {
search?: string;
page?: number;
limit?: number;
};