# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Vigyanics API — public auth + admin panel
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/auth/login:
    post:
      operationId: loginAdmin
      tags: [admin-auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/create-admin:
    post:
      operationId: createAdmin
      tags: [admin-auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [admin-categories]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [admin-categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [admin-categories]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [admin-categories]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin-orders]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /admin/orders/{id}:
    get:
      operationId: getAdminOrder
      tags: [admin-orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin-orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin-customers]
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
  /admin/customers/{id}/status:
    patch:
      operationId: updateCustomerStatus
      tags: [admin-customers]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    CreateAdminInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Category:
      type: object
      required: [id, name, slug, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt, updatedAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
        total:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [id, orderId, productName, quantity, price, total]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: string
        total:
          type: string
    OrderDetail:
      type: object
      required: [id, orderNumber, status, total, items, createdAt, updatedAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
        total:
          type: string
        notes:
          type: ["string", "null"]
        shippingAddress:
          type: ["object", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    CustomerListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CustomerStatusUpdate:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean