# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vigyanics API — public auth + admin panel
 * OpenAPI spec version: 0.2.0
 */

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
}