# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vigyanics API — public auth + admin panel
 * OpenAPI spec version: 0.2.0
 */

export interface Order {
  id: number;
  orderNumber: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  status: string;
  total: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}