# lib/api-zod/src/generated/types/orderDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vigyanics API — public auth + admin panel
 * OpenAPI spec version: 0.2.0
 */
import type { OrderDetailShippingAddress } from './orderDetailShippingAddress';
import type { OrderItem } from './orderItem';

export interface OrderDetail {
  id: number;
  orderNumber: string;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  status: string;
  total: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  shippingAddress?: OrderDetailShippingAddress;
  items: OrderItem[];
  createdAt: Date;
  updatedAt: Date;
}